no idea what im doing, but changed gevolutionlines to u32 instead of u16

+++ include/pokemon.h
@@ -264,7 +264,7 @@ extern const u8 gStatStageRatios[][2];
 extern const u16 gLinkPlayerFacilityClasses[];
 extern const struct SpriteTemplate gUnknown_08329D98[];
 extern const s8 gNatureStatTable[][5];
-extern const u16 gEvolutionLines[NUM_SPECIES][EVOS_PER_LINE];
+extern const u32 gEvolutionLines[NUM_SPECIES][EVOS_PER_LINE];
 
 void ZeroBoxMonData(struct BoxPokemon *boxMon);
 void ZeroMonData(struct Pokemon *mon);
